Working on multiboot. Improve boot.s
