+++ src/include/cpu.hpp
@@ -48,12 +48,12 @@
 #define CR0_TS       0x00000008
 #define CR0_ET       0x00000010
 
-// NE：对于Intel 80486或以上的CPU，CR0的位5是协处理器错误（Numeric Error）标志。
-// 当设置该标志时，就启用了x87协处理器错误的内部报告机制；若复位该标志，那么就使用PC形式的x87协处理器错误报告机制。
-// 当NE为复位状态并且CPU的IGNNE输入引脚有信号时，那么数学协处理器x87错误将被忽略。
-// 当NE为复位状态并且CPU的IGNNE输入引脚无信号时，那么非屏蔽的数学协处理器x87错误将导致处理器通过FERR引脚在外部产生一个中断，
-// 并且在执行下一个等待形式浮点指令或WAIT/FWAIT指令之前立刻停止指令执行。CPU的FERR引脚用于仿真外部协处理器80387的ERROR引脚，
-// 因此通常连接到中断控制器输入请求引脚上。NE标志、IGNNE引脚和FERR引脚用于利用外部逻辑来实现PC形式的外部错误报告机制。
+// NE：对于 Intel 80486或以上的CPU，CR0 的位5是协处理器错误（Numeric Error）标志。
+// 当设置该标志时，就启用了 x87 协处理器错误的内部报告机制；若复位该标志，那么就使用 PC 形式的 x87 协处理器错误报告机制。
+// 当NE为复位状态并且 CPU 的 IGNNE 输入引脚有信号时，那么数学协处理器 x87 错误将被忽略。
+// 当NE为复位状态并且 CPU 的 IGNNE 输入引脚无信号时，那么非屏蔽的数学协处理器 x87 错误将导致处理器通过 FERR 引脚在外部产生一个中断，
+// 并且在执行下一个等待形式浮点指令或 WAIT/FWAIT 指令之前立刻停止指令执行。 CPU 的 FERR 引脚用于仿真外部协处理器 80387 的 ERROR 引脚，
+// 因此通常连接到中断控制器输入请求引脚上。NE 标志、IGNNE 引脚和 FERR 引脚用于利用外部逻辑来实现 PC 形式的外部错误报告机制。
 #define CR0_NE       0x00000020
 
 // WP：对于Intel 80486或以上的CPU，CR0的位16是写保护（Write Proctect）标志。
